Version 2.30
CHG: For :FilePlus, :WritePlus, :SavePlus, when a [count] but no [!] is given, try to create an offset one more than an existing file between the current and the passed offset. This lets you use a large [N] to write the file with the next number within [N] for which no file exists yet. Change inspired by
http://stackoverflow.com/questions/13778322/vimscript-code-to-create-a-new-numbered-file